Start breaking up the codegen backend.
