+++ src/python/pants/backend/codegen/subsystems/BUILD
@@ -1,11 +1,9 @@
-# Copyright 2015 Pants project contributors (see CONTRIBUTORS.md).
+# Copyright 2016 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
 python_library(
   name='thrift_defaults',
-  sources=['thrift_defaults.py'],
   dependencies=[
-    'src/python/pants/backend/codegen/targets:java',
-    'src/python/pants/subsystem',
+    'src/python/pants/backend/codegen/thrift/java',
   ],
 )
