+++ src/python/pants/backend/codegen/thrift/java/BUILD
@@ -1,0 +1,19 @@
+# Copyright 2016 Pants project contributors (see CONTRIBUTORS.md).
+# Licensed under the Apache License, Version 2.0 (see LICENSE).
+
+python_library(
+  dependencies = [
+    '3rdparty/python/twitter/commons:twitter.common.collections',
+    'src/python/pants/backend/jvm/targets:java',
+    'src/python/pants/base:build_environment',
+    'src/python/pants/base:exceptions',
+    'src/python/pants/base:workunit',
+    'src/python/pants/binaries:thrift_util',
+    'src/python/pants/build_graph',
+    'src/python/pants/goal:task_registrar',
+    'src/python/pants/option',
+    'src/python/pants/subsystem',
+    'src/python/pants/task',
+    'src/python/pants/util:memo',
+  ],
+)
