+++ src/python/pants/backend/codegen/thrift/java/register.py
@@ -1,0 +1,23 @@
+# coding=utf-8
+# Copyright 2014 Pants project contributors (see CONTRIBUTORS.md).
+# Licensed under the Apache License, Version 2.0 (see LICENSE).
+
+from __future__ import (absolute_import, division, generators, nested_scopes, print_function,
+                        unicode_literals, with_statement)
+
+from pants.backend.codegen.thrift.java.apache_thrift_gen import ApacheThriftGen
+from pants.backend.codegen.thrift.java.java_thrift_library import JavaThriftLibrary
+from pants.build_graph.build_file_aliases import BuildFileAliases
+from pants.goal.task_registrar import TaskRegistrar as task
+
+
+def build_file_aliases():
+  return BuildFileAliases(
+    targets={
+      'java_thrift_library': JavaThriftLibrary,
+      }
+    )
+
+
+def register_goals():
+  task(name='thrift', action=ApacheThriftGen).install('gen')
