+++ src/python/pants/backend/python/BUILD
@@ -116,6 +116,7 @@ python_library(
     ':python_requirement',
     ':thrift_builder',
     'src/python/pants/backend/codegen/targets:python',
+    'src/python/pants/backend/codegen/thrift/python',
     'src/python/pants/backend/python/targets:python',
     'src/python/pants/base:build_environment',
     'src/python/pants/build_graph',
@@ -150,11 +151,11 @@ python_library(
     ':code_generator',
       # Note that this is part of a gnarly dependency chain:
       #  backend/python:thrift_builder ->
-      #  backend/codegen/targets:python ->
+      #  src/python/pants/backend/codegen/thrift/python ->
       #  backend/python/targets:python
       # This isn't circular, but it's still bad that it crosses between backends and back again.
       # TODO: Clean this up. thrift_builder should almost certainly be in the codegen backend.
-    'src/python/pants/backend/codegen/targets:python',
+    'src/python/pants/backend/codegen/thrift/python',
     'src/python/pants/base:build_environment',
     'src/python/pants/util:dirutil',
     'src/python/pants/util:memo',
