+++ tests/python/pants_test/backend/codegen/thrift/java/BUILD
@@ -2,12 +2,12 @@
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
 python_tests(
-  name='thrift_defaults',
-  sources=['test_thrift_defaults.py'],
   dependencies=[
-    'src/python/pants/backend/codegen/subsystems:thrift_defaults',
-    'src/python/pants/backend/codegen/targets:java',
+    'src/python/pants/backend/codegen/thrift/java',
+    'src/python/pants/backend/jvm/targets:java',
+    'src/python/pants/base:exceptions',
     'src/python/pants/build_graph',
     'tests/python/pants_test:base_test',
+    'tests/python/pants_test/tasks:task_test_base',
   ]
 )
