+++ tests/python/pants_test/backend/python/BUILD
@@ -19,6 +19,7 @@ python_tests(
   dependencies=[
     '3rdparty/python:pex',
     'src/python/pants/backend/codegen/targets:python',
+    'src/python/pants/backend/codegen/thrift/python',
     'src/python/pants/backend/python/targets:python',
     'src/python/pants/backend/python:interpreter_cache',
     'src/python/pants/backend/python:python_chroot',
